requirements.txt
autopep8==2.3.2
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
google-api-core==2.25.1
google-api-python-client==2.181.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
gspread==6.2.1
gspread-formatting==1.2.1
httplib2==0.30.0
idna==3.10
markdown-it-py==4.0.0
mdurl==0.1.2
oauthlib==3.3.1
prompt_toolkit==3.0.52
proto-plus==1.26.1
protobuf==6.32.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.14.0
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
questionary==2.1.1
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.1.0
rsa==4.9.1
six==1.17.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
wcwidth==0.2.13